+++ fl_sim/dataset/datasets/emnist.py
@@ -1,7 +1,6 @@
 import tensorflow as tf
 import numpy as np
 import tensorflow_datasets.public_api as tfds
-from sklearn.model_selection import train_test_split
 from ..model_loader import DatasetModelLoader
 
 
@@ -31,20 +30,20 @@ class Emnist(DatasetModelLoader):
         return x_train, y_train, x_test, y_test
 
     # Image classification task
-    def get_compiled_model(self, optimizer: str, metric: str, train_data):
-        model = tf.keras.models.Sequential()
-        model.add(tf.keras.layers.Conv2D(32, kernel_size=(3, 3), strides=1, activation='relu', input_shape=(28, 28, 1)))
-        model.add(tf.keras.layers.MaxPooling2D((2, 2)))
-        model.add(tf.keras.layers.Conv2D(32, (3, 3), activation='relu', strides=1))
-        model.add(tf.keras.layers.MaxPooling2D((2, 2)))
-        model.add(tf.keras.layers.Flatten())
-        model.add(tf.keras.layers.Dense(64, activation='relu'))
-        model.add(tf.keras.layers.Dense(62, activation='softmax'))
+    def get_compiled_model(self, optimizer: str, metric: str, train_data):  # https://www.tensorflow.org/tutorials/quickstart/beginner
+        tf_model = tf.keras.models.Sequential(
+            [
+                tf.keras.layers.Flatten(input_shape=(28, 28)),
+                tf.keras.layers.Dense(128, activation="relu"),
+                tf.keras.layers.Dropout(0.2),
+                tf.keras.layers.Dense(62, activation="softmax"),
+            ]
+        )
 
-        model.compile(
+        tf_model.compile(
             optimizer=optimizer, loss="sparse_categorical_crossentropy", metrics=[metric]
         )
-        return model
+        return tf_model
 
     def get_loss_function(self):
         return "sparse_categorical_crossentropy"