+++ fl_sim/dataset/datasets/tff_emnist.py
@@ -26,19 +26,19 @@ class Emnist_tff(DatasetModelLoader):
         tf_model = tf.keras.models.Sequential(
             [
                 tf.keras.layers.Flatten(input_shape=(28, 28)),
-                tf.keras.layers.Dense(10, kernel_initializer='zeros'),
-                tf.keras.layers.Softmax(),
+                tf.keras.layers.Dense(128, activation="relu"),
+                tf.keras.layers.Dropout(0.2),
+                tf.keras.layers.Dense(62, activation="softmax"),
             ]
         )
 
         tf_model.compile(
-            optimizer=optimizer, loss=tf.keras.losses.SparseCategoricalCrossentropy(), metrics=[metric]
+            optimizer=optimizer, loss="sparse_categorical_crossentropy", metrics=[metric]
         )
         return tf_model
 
     def get_loss_function(self):
         return tf.keras.losses.SparseCategoricalCrossentropy()
-
 
     @staticmethod
     def get_images_and_labels(federated_data):